// SETPPBO/ClientApp/app/main/penetapan/models/penetapan.ts
export class Penetapan {
    PermohonanId:number;
    NoPenetapan: string;    
    TglPenetapan: Date;
    TempatSidang: number;
    PejabatPenandatangan: string;
    PosisiPejabatPenandatangan: string;
    kotaId: string;
    kotapenetapan: string;
    RefJenisPermohonanId: number;
    NamaTermohon: string;
    NamaUpTermohon: string;

}